Major project restructurization + optimization
