--- ClickHaskell/ClickHaskell/NativeProtocol/ClientPackets.hs
@@ -1,258 +1,0 @@
-{-# LANGUAGE
-    DuplicateRecordFields
-  , OverloadedStrings
-#-}
-
-module ClickHaskell.NativeProtocol.ClientPackets where
-
--- Internal dependencies
-import ClickHaskell.DbTypes
-import ClickHaskell.NativeProtocol.Serialization
-import ClickHaskell.NativeProtocol.Columns (Columns, KnownColumns(columnsCount, rowsCount))
-
--- GHC included
-import Data.Text (Text)
-import Data.Typeable (Proxy (..))
-import GHC.Generics
-import GHC.TypeLits (KnownNat, natVal)
-
--- * Client packets
-
-data ClientPacketType
-  = Hello
-  | Query
-  | Data
-  | Cancel
-  | Ping
-  | TablesStatusRequest
-  | KeepAlive
-  | Scalar
-  | IgnoredPartUUIDs
-  | ReadTaskResponse
-  | MergeTreeReadTaskResponse
-  | SSHChallengeRequest
-  | SSHChallengeResponse
-  deriving (Enum, Show)
-
-type family PacketTypeNumber (packetType :: ClientPacketType)
-  where
-  PacketTypeNumber Hello = 0
-  PacketTypeNumber Query = 1
-  PacketTypeNumber Data = 2
-  PacketTypeNumber Cancel = 3
-  PacketTypeNumber Ping = 4
-  PacketTypeNumber TablesStatusRequest = 5
-  PacketTypeNumber KeepAlive = 6
-  PacketTypeNumber Scalar = 7
-  PacketTypeNumber IgnoredPartUUIDs = 8
-  PacketTypeNumber ReadTaskResponse = 9
-  PacketTypeNumber MergeTreeReadTaskResponse = 10
-  PacketTypeNumber SSHChallengeRequest = 11
-  PacketTypeNumber SSHChallengeResponse = 12
-
-data Packet (packetType :: ClientPacketType) = MkPacket
-instance KnownNat (PacketTypeNumber packetType) => Show (Packet (packetType :: ClientPacketType)) where
-  show _ = show . toEnum @ClientPacketType . fromIntegral $ packetNumVal @packetType
-
-packetNumVal :: forall packetType . KnownNat (PacketTypeNumber packetType) => UVarInt
-packetNumVal = fromIntegral . natVal $ Proxy @(PacketTypeNumber packetType)
-
-instance
-  KnownNat (PacketTypeNumber packetType)
-  =>
-  Serializable (Packet (packetType :: ClientPacketType)) where
-  serialize rev _ = serialize @UVarInt rev (packetNumVal @packetType)
-
-instance Deserializable (Packet (packetType :: ClientPacketType)) where
-  deserialize _rev = pure (MkPacket @packetType)
-
-
--- ** Hello
-
-data HelloParameters = MkHelloParameters
-  { chDatabase :: Text
-  , chLogin :: Text
-  , chPass :: Text
-  }
-
-mkHelloPacket :: HelloParameters -> HelloPacket
-mkHelloPacket MkHelloParameters{chDatabase, chLogin, chPass} =
-  MkHelloPacket
-    { packet_type          = MkPacket
-    , client_name          = clientNameAndVersion
-    , client_version_major = clientMajorVersion
-    , client_version_minor = clientMinorVersion
-    , tcp_protocol_version = latestSupportedRevision
-    , default_database     = toChType chDatabase
-    , user                 = toChType chLogin
-    , password             = toChType chPass
-    }
-
-data HelloPacket = MkHelloPacket
-  { packet_type          :: Packet Hello
-  , client_name          :: ChString
-  , client_version_major :: UVarInt
-  , client_version_minor :: UVarInt
-  , tcp_protocol_version :: UVarInt
-  , default_database     :: ChString
-  , user                 :: ChString
-  , password             :: ChString
-  }
-  deriving (Generic, Serializable)
-
-
-mkAddendum :: Addendum
-mkAddendum = MkAddendum
-  { quota_key = MkSinceRevision ""
-  }
-
-data Addendum = MkAddendum
-  { quota_key :: ChString `SinceRevision` DBMS_MIN_PROTOCOL_VERSION_WITH_QUOTA_KEY
-  }
-  deriving (Generic, Serializable)
-
-
--- ** Ping
-
-mkPingPacket :: PingPacket
-mkPingPacket = MkPingPacket{packet_type = MkPacket}
-
-data PingPacket = MkPingPacket{packet_type :: Packet Ping}
-  deriving (Generic, Serializable)
-
-
--- ** Query
-
-mkQueryPacket :: ProtocolRevision -> ChString -> ChString -> QueryPacket
-mkQueryPacket chosenRev user query = MkQueryPacket
-  { query_packet = MkPacket
-  , query_id = ""
-  , client_info                    = MkSinceRevision MkClientInfo
-    { query_kind                   = InitialQuery
-    , initial_user                 = user
-    , initial_query_id             = ""
-    , initial_adress               = "0.0.0.0:0"
-    , initial_time                 = MkSinceRevision 0
-    , interface_type               = 1 -- [tcp - 1, http - 2]
-    , os_user                      = "dmitry"
-    , hostname                     = "desktop"
-    , client_name                  = clientNameAndVersion
-    , client_major                 = clientMajorVersion
-    , client_minor                 = clientMinorVersion
-    , client_revision              = chosenRev
-    , quota_key                    = MkSinceRevision ""
-    , distrubuted_depth            = MkSinceRevision 0
-    , client_patch                 = MkSinceRevision clientPatchVersion
-    , open_telemetry               = MkSinceRevision 0
-    , collaborate_with_initiator   = MkSinceRevision 0
-    , count_participating_replicas = MkSinceRevision 0
-    , number_of_current_replica    = MkSinceRevision 0
-    }
-  , settings           = MkDbSettings
-  , interserver_secret = MkSinceRevision ""
-  , query_stage        = Complete
-  , compression        = 0
-  , query              = query
-  , parameters         = MkSinceRevision MkQueryParameters
-  }
-
-data QueryPacket = MkQueryPacket
-  { query_packet       :: Packet Query
-  , query_id           :: ChString
-  , client_info        :: ClientInfo `SinceRevision` DBMS_MIN_REVISION_WITH_CLIENT_INFO
-  , settings           :: DbSettings
-  , interserver_secret :: ChString `SinceRevision` DBMS_MIN_REVISION_WITH_INTERSERVER_SECRET
-  , query_stage        :: QueryStage
-  , compression        :: UVarInt
-  , query              :: ChString
-  , parameters         :: QueryParameters `SinceRevision` DBMS_MIN_PROTOCOL_VERSION_WITH_PARAMETERS
-  }
-  deriving (Generic, Serializable)
-
-data DbSettings = MkDbSettings
-instance Serializable DbSettings where serialize rev _ = serialize @ChString rev ""
-
-data QueryParameters = MkQueryParameters
-instance Serializable QueryParameters where serialize rev _ = serialize @ChString rev ""
-
-data QueryStage
-  = FetchColumns
-  | WithMergeableState
-  | Complete
-  | WithMergeableStateAfterAggregation
-  | WithMergeableStateAfterAggregationAndLimit
-  deriving (Enum)
-
-instance Serializable QueryStage where
-  serialize rev = serialize @UVarInt rev . fromIntegral . fromEnum
-
-queryStageCode :: QueryStage -> UVarInt
-queryStageCode = fromIntegral . fromEnum
-
-data Flags = IMPORTANT | CUSTOM | OBSOLETE
-flagCode :: Flags -> ChUInt64
-flagCode IMPORTANT = 0x01
-flagCode CUSTOM    = 0x02
-flagCode OBSOLETE  = 0x04
-
-data ClientInfo = MkClientInfo
-  { query_kind                   :: QueryKind
-  , initial_user                 :: ChString
-  , initial_query_id             :: ChString
-  , initial_adress               :: ChString
-  , initial_time                 :: ChInt64 `SinceRevision` DBMS_MIN_PROTOCOL_VERSION_WITH_INITIAL_QUERY_START_TIME
-  , interface_type               :: ChUInt8
-  , os_user                      :: ChString
-  , hostname                     :: ChString
-  , client_name                  :: ChString
-  , client_major                 :: UVarInt
-  , client_minor                 :: UVarInt
-  , client_revision              :: UVarInt
-  , quota_key                    :: ChString `SinceRevision` DBMS_MIN_REVISION_WITH_QUOTA_KEY_IN_CLIENT_INFO
-  , distrubuted_depth            :: UVarInt `SinceRevision` DBMS_MIN_PROTOCOL_VERSION_WITH_DISTRIBUTED_DEPTH
-  , client_patch                 :: UVarInt `SinceRevision` DBMS_MIN_REVISION_WITH_VERSION_PATCH
-  , open_telemetry               :: ChUInt8 `SinceRevision` DBMS_MIN_REVISION_WITH_OPENTELEMETRY
-  , collaborate_with_initiator   :: UVarInt `SinceRevision` DBMS_MIN_REVISION_WITH_PARALLEL_REPLICAS
-  , count_participating_replicas :: UVarInt `SinceRevision` DBMS_MIN_REVISION_WITH_PARALLEL_REPLICAS
-  , number_of_current_replica    :: UVarInt `SinceRevision` DBMS_MIN_REVISION_WITH_PARALLEL_REPLICAS
-  }
-  deriving (Generic, Serializable)
-
-data QueryKind = NoQuery | InitialQuery | SecondaryQuery
-  deriving (Enum)
-
-instance Serializable QueryKind where
-  serialize rev = serialize @ChUInt8 rev . fromIntegral . fromEnum
-
-
--- ** Data
-
-mkDataPacket :: forall columns . KnownColumns (Columns columns) => ChString -> Columns columns -> DataPacket
-mkDataPacket table_name columns =
-  MkDataPacket
-    { packet_type   = MkPacket
-    , table_name
-    , block_info    = MkBlockInfo
-      { field_num1   = 1, is_overflows = 0
-      , field_num2   = 2, bucket_num   = -1
-      , eof          = 0
-      }
-    , columns_count = columnsCount @(Columns columns)
-    , rows_count    = rowsCount columns
-    }
-
-data DataPacket = MkDataPacket
-  { packet_type   :: Packet Data
-  , table_name    :: ChString
-  , block_info    :: BlockInfo
-  , columns_count :: UVarInt
-  , rows_count    :: UVarInt
-  }
-  deriving (Generic, Serializable, Deserializable, Show)
-
-data BlockInfo = MkBlockInfo
-  { field_num1   :: UVarInt, is_overflows :: ChUInt8
-  , field_num2   :: UVarInt, bucket_num   :: ChInt32
-  , eof          :: UVarInt
-  }
-  deriving (Generic, Serializable, Deserializable, Show)
