--- ClickHaskell/ClickHaskell/NativeProtocol/ServerPackets.hs
@@ -1,139 +1,0 @@
-{-# LANGUAGE
-    DuplicateRecordFields
-  , RankNTypes
-#-}
-
-module ClickHaskell.NativeProtocol.ServerPackets where
-
--- Internal dependencies
-import ClickHaskell.DbTypes
-import ClickHaskell.NativeProtocol.Serialization
-import ClickHaskell.NativeProtocol.ClientPackets (DataPacket)
-
--- GHC included
-import GHC.Generics (Generic)
-
--- * Server packets
-
-data ServerPacketType where
-  HelloResponse :: HelloResponse -> ServerPacketType
-  DataResponse :: DataPacket -> ServerPacketType
-  Exception :: ExceptionPacket -> ServerPacketType
-  Progress :: ProgressPacket -> ServerPacketType
-  Pong :: ServerPacketType
-  EndOfStream :: ServerPacketType
-  ProfileInfo :: ProfileInfo -> ServerPacketType
-  Totals :: ServerPacketType
-  Extremes :: ServerPacketType
-  TablesStatusResponse :: ServerPacketType
-  Log :: ServerPacketType
-  TableColumns :: ServerPacketType
-  UUIDs :: ServerPacketType
-  ReadTaskRequest :: ServerPacketType
-  ProfileEvents :: ServerPacketType
-  UnknownPacket :: UVarInt -> ServerPacketType
-
-instance Deserializable ServerPacketType where
-  deserialize rev = do
-    packetNum <- deserialize @UVarInt rev
-    case packetNum of
-      0  -> HelloResponse <$> deserialize rev
-      1  -> DataResponse <$> deserialize rev
-      2  -> Exception <$> deserialize rev
-      3  -> Progress <$> deserialize rev
-      4  -> pure Pong
-      5  -> pure EndOfStream
-      6  -> ProfileInfo <$> deserialize rev
-      7  -> pure Totals
-      8  -> pure Extremes
-      9  -> pure TablesStatusResponse
-      10 -> pure Log
-      11 -> pure TableColumns
-      12 -> pure UUIDs
-      13 -> pure ReadTaskRequest
-      14 -> pure ProfileEvents
-      _  -> pure $ UnknownPacket packetNum
-
-instance Show ServerPacketType where
-  show (HelloResponse hello) = "HelloResponse " <> show hello
-  show (DataResponse dataPacket) = "DataResponse" <> show dataPacket
-  show (Exception exception) = "Exception" <> show exception
-  show (Progress progress) = "Progress" <> show progress
-  show Pong = "Pong"
-  show EndOfStream = "EndOfStream"
-  show (ProfileInfo profileInfo) = "ProfileInfo" <> show profileInfo 
-  show Totals = "Totals"
-  show Extremes = "Extremes"
-  show TablesStatusResponse = "TablesStatusResponse"
-  show Log = "Log"
-  show TableColumns = "TableColumns"
-  show UUIDs = "UUIDs"
-  show ReadTaskRequest = "ReadTaskRequest"
-  show ProfileEvents = "ProfileEvents"
-  show (UnknownPacket packetNum) = "UnknownPacket: " <> show packetNum
-
--- ** HelloResponse
-
-{-
-  https://github.com/ClickHouse/ClickHouse/blob/eb4a74d7412a1fcf52727cd8b00b365d6b9ed86c/src/Client/Connection.cpp#L520
--}
-data HelloResponse = MkHelloResponse
-  { server_name                    :: ChString
-  , server_version_major           :: UVarInt
-  , server_version_minor           :: UVarInt
-  , server_revision                :: UVarInt
-  , server_parallel_replicas_proto :: UVarInt  `SinceRevision` DBMS_MIN_REVISION_WITH_VERSIONED_PARALLEL_REPLICAS_PROTOCOL
-  , server_timezone                :: ChString `SinceRevision` DBMS_MIN_REVISION_WITH_SERVER_TIMEZONE
-  , server_display_name            :: ChString `SinceRevision` DBMS_MIN_REVISION_WITH_SERVER_DISPLAY_NAME
-  , server_version_patch           :: UVarInt  `SinceRevision` DBMS_MIN_REVISION_WITH_VERSION_PATCH
-  , proto_send_chunked_srv         :: ChString `SinceRevision` DBMS_MIN_PROTOCOL_VERSION_WITH_CHUNKED_PACKETS
-  , proto_recv_chunked_srv         :: ChString `SinceRevision` DBMS_MIN_PROTOCOL_VERSION_WITH_CHUNKED_PACKETS
-  , password_complexity_rules      :: [PasswordComplexityRules] `SinceRevision` DBMS_MIN_PROTOCOL_VERSION_WITH_PASSWORD_COMPLEXITY_RULES
-  , read_nonce                     :: ChUInt64 `SinceRevision` DBMS_MIN_REVISION_WITH_INTERSERVER_SECRET_V2
-  }
-  deriving (Generic, Deserializable, Show)
-
-data PasswordComplexityRules = MkPasswordComplexityRules
-  { original_pattern  :: ChString
-  , exception_message :: ChString
-  }
-  deriving (Generic, Deserializable, Show)
-
--- ** Exception
-
-data ExceptionPacket = MkExceptionPacket
-  { code        :: ChInt32
-  , name        :: ChString
-  , message     :: ChString
-  , stack_trace :: ChString
-  , nested      :: ChUInt8
-  }
-  deriving (Generic, Deserializable, Show)
-
-
--- ** Progress
-
-data ProgressPacket = MkProgressPacket
-  { rows        :: UVarInt
-  , bytes       :: UVarInt
-  , total_rows  :: UVarInt
-  , total_bytes :: UVarInt `SinceRevision` DBMS_MIN_PROTOCOL_VERSION_WITH_TOTAL_BYTES_IN_PROGRESS
-  , wrote_rows  :: UVarInt `SinceRevision` DBMS_MIN_PROTOCOL_VERSION_WITH_TOTAL_BYTES_IN_PROGRESS
-  , wrote_bytes :: UVarInt `SinceRevision` DBMS_MIN_REVISION_WITH_CLIENT_WRITE_INFO
-  , elapsed_ns  :: UVarInt `SinceRevision` DBMS_MIN_REVISION_WITH_CLIENT_WRITE_INFO
-  }
-  deriving (Generic, Deserializable, Show)
-
--- ** ProfileInfo
-
-data ProfileInfo = MkProfileInfo
-  { rows                         :: UVarInt
-  , blocks                       :: UVarInt
-  , bytes                        :: UVarInt
-  , applied_limit                :: ChUInt8
-  , rows_before_limit            :: UVarInt
-  , calculated_rows_before_limit :: ChUInt8
-  , applied_aggregation          :: ChUInt8 `SinceRevision` DBMS_MIN_REVISION_WITH_ROWS_BEFORE_AGGREGATION
-  , rows_before_aggregation      :: UVarInt `SinceRevision` DBMS_MIN_REVISION_WITH_ROWS_BEFORE_AGGREGATION
-  }
-  deriving (Generic, Deserializable, Show)
